fixs

+++ server/src/middlewares/image.upload.middleware.ts
@@ -3,14 +3,19 @@ import sharp from 'sharp'
 import { middlewareInterface } from '../types/middleware.interface';
 import { generateFileName } from '../helpers/upload.helper';
 import { uploadFile, deleteFile, getObjectSignedUrl } from '../services/awsS3'
+import { ErrorHelper } from '../helpers/error.helper';
 
 //  multer
 const storage = multer.memoryStorage()
 const uploadMulter = multer({ storage: storage });
 
 // add image
-const uploadImage: middlewareInterface['uploadImage'] = async (resizeOptions, req, res, next) => {
+const uploadImage: middlewareInterface['uploadImage'] = async (resizeOptions = { height: 520, width: 520 }, req, res, next, update) => {
+    if (update && !req.body.image_name) {
+        return next()
+    }
     const file = req.file
+
     if (file) {
         const imageName: string = req.body.image_name || generateFileName()
 
@@ -21,22 +26,24 @@ const uploadImage: middlewareInterface['uploadImage'] = async (resizeOptions, re
         await uploadFile(fileBuffer, imageName, file.mimetype)
 
         if (!req.body.image_name) {
-            const image_url = getObjectSignedUrl(imageName)
-            req.body.image_name = imageName
-            req.body.image_url = image_url
+            const image_url = await getObjectSignedUrl(imageName)
+            res.locals.image_name = imageName
+            res.locals.image_url = image_url
         }
     }
-    next()
+    return next()
 
 }
 
 const deleteImage: middlewareInterface['deleteImage'] = async (req, res, next) => {
 
     const imageName: string = req.body.image_name
-
+    if (!imageName) {
+        throw new ErrorHelper('Try again!')
+    }
     await deleteFile(imageName)
 
-    next()
+    return next()
 
 }
